+++ test/issues-test.js
@@ -54,36 +54,43 @@ describe('issues', () => {
         expect(children.map(({ id }) => id)).to.eql(['Start', 'Parallel1', 'Task_A', 'Task_B', 'Parallel2', 'Task_C', 'End']);
       }
 
+      // eslint-disable-next-line no-useless-assignment
       let [Start, Parallel1, Task_A, Task_B, Parallel2, Task_C, End] = states[0].definitions[0].execution.processes[0].execution.children;
       expect(Start, 'state 0 Start').to.have.property('status', 'started');
       expect(Parallel1, 'state 0 Parallel1').to.not.have.property('status');
 
+      // eslint-disable-next-line no-useless-assignment
       [Start, Parallel1, Task_A, Task_B, Parallel2, Task_C, End] = states[1].definitions[0].execution.processes[0].execution.children;
       expect(Start, 'state 1 Start').to.have.property('status', 'end');
       expect(Parallel1, 'state 1 Parallel1').to.have.property('status', 'started');
       expect(Task_A, 'state 1 Task_A').to.not.have.property('status');
 
+      // eslint-disable-next-line no-useless-assignment
       [Start, Parallel1, Task_A, Task_B, Parallel2, Task_C, End] = states[2].definitions[0].execution.processes[0].execution.children;
       expect(Parallel1, 'state 2 Parallel1').to.have.property('status', 'end');
       expect(Task_A, 'state 2 Task_A').to.have.property('status', 'started');
       expect(Task_B, 'state 2 Task_B').to.not.have.property('status');
 
+      // eslint-disable-next-line no-useless-assignment
       [Start, Parallel1, Task_A, Task_B, Parallel2, Task_C, End] = states[3].definitions[0].execution.processes[0].execution.children;
       expect(Parallel1, 'state 3 Parallel1').to.have.property('status', 'end');
-      expect(Task_A, 'state 3 Task_A').to.not.have.property('status');
-      expect(Task_B, 'state 3 Task_B').to.have.property('status', 'started');
-      expect(Parallel2, 'state 3 Parallel2').to.not.have.property('status');
+      expect(Task_A, 'state 3 Task_A').to.have.property('status', 'end');
+      expect(Task_B, 'state 3 Task_B').to.not.have.property('status');
+      expect(Parallel2, 'state 3 Parallel2').to.have.property('status', 'started');
 
+      // eslint-disable-next-line no-useless-assignment
       [Start, Parallel1, Task_A, Task_B, Parallel2, Task_C, End] = states[4].definitions[0].execution.processes[0].execution.children;
       expect(Task_A, 'state 4 Task_A').to.not.have.property('status');
-      expect(Task_B, 'state 4 Task_B').to.have.property('status', 'end');
-      expect(Parallel2, 'state 4 Parallel2').to.have.property('status', 'started');
+      expect(Task_B, 'state 4 Task_B').to.have.property('status', 'started');
+      expect(Parallel2, 'state 4 Parallel2').to.have.property('status', 'executing');
 
+      // eslint-disable-next-line no-useless-assignment
       [Start, Parallel1, Task_A, Task_B, Parallel2, Task_C, End] = states[5].definitions[0].execution.processes[0].execution.children;
       expect(Parallel2, 'state 5 Parallel2').to.have.property('status', 'end');
       expect(Task_C, 'state 5 Task_C').to.have.property('status', 'started');
       expect(End, 'state 5 End').to.not.have.property('status');
 
+      // eslint-disable-next-line no-useless-assignment
       [Start, Parallel1, Task_A, Task_B, Parallel2, Task_C, End] = states[6].definitions[0].execution.processes[0].execution.children;
       expect(Task_C, 'state 6 Task_C').to.have.property('status', 'end');
       expect(End, 'state 6 End').to.have.property('status', 'started');